+++ changelog.md
@@ -18,6 +18,7 @@
 + Removed custom logic for local shops.
 + Added sentry validation to avoid crash if sentry breaks down.
 + Added check to avoid exception in path utility class.
++ Using single click payment without component now ask if you want to save it or use saved card.
 + Fixed issue where klarna status would be changed to klarna shipped and not completed when using default statuses for klarna.
 
 ## Changes in release 4.4.3 ##